# k-vet-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "k-vet API",
    "description": "Veterinary practice management — single user, self-hosted.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "operationId": "patchSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "422": {
            "description": "An address is malformed, or the logo does not exist"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatchSettings": {
        "type": "object",
        "properties": {
          "practice_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "practice_street": {
            "type": [
              "string",
              "null"
            ]
          },
          "practice_zip": {
            "type": [
              "string",
              "null"
            ]
          },
          "practice_city": {
            "type": [
              "string",
              "null"
            ]
          },
          "practice_country": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "iban": {
            "type": [
              "string",
              "null"
            ]
          },
          "bic": {
            "type": [
              "string",
              "null"
            ]
          },
          "bank_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "ustid": {
            "type": [
              "string",
              "null"
            ]
          },
          "logo_attachment_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "cc_emails": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "bcc_emails": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "required": [
          "practice_name",
          "practice_street",
          "practice_zip",
          "practice_city",
          "email",
          "iban",
          "bic",
          "bank_name",
          "ustid",
          "cc_emails",
          "bcc_emails"
        ],
        "properties": {
          "practice_name": {
            "type": "string"
          },
          "practice_street": {
            "type": "string"
          },
          "practice_zip": {
            "type": "string"
          },
          "practice_city": {
            "type": "string"
          },
          "practice_country": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 3166-1 alpha-2; empty falls back to `[invoice] default_country`."
          },
          "email": {
            "type": "string",
            "description": "Shown in the invoice footer. The sender address of outgoing mail is operator\nconfiguration and lives in `config.toml`."
          },
          "iban": {
            "type": "string"
          },
          "bic": {
            "type": "string"
          },
          "bank_name": {
            "type": "string"
          },
          "ustid": {
            "type": "string"
          },
          "logo_attachment_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "cc_emails": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Every outgoing invoice email is copied to these addresses."
          },
          "bcc_emails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}