# k-vet-web/src/api/generated/model/patchSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * k-vet API
 * Veterinary practice management — single user, self-hosted.
 * OpenAPI spec version: 0.1.0
 */

export interface PatchSettings {
  /** @nullable */
  practice_name?: string | null;
  /** @nullable */
  practice_street?: string | null;
  /** @nullable */
  practice_zip?: string | null;
  /** @nullable */
  practice_city?: string | null;
  /** @nullable */
  practice_country?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  iban?: string | null;
  /** @nullable */
  bic?: string | null;
  /** @nullable */
  bank_name?: string | null;
  /** @nullable */
  ustid?: string | null;
  /** @nullable */
  logo_attachment_id?: number | null;
  /** @nullable */
  cc_emails?: string[] | null;
  /** @nullable */
  bcc_emails?: string[] | null;
}