# k-vet-web/src/api/generated/model/settings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * k-vet API
 * Veterinary practice management — single user, self-hosted.
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  practice_name: string;
  practice_street: string;
  practice_zip: string;
  practice_city: string;
  /**
     * ISO 3166-1 alpha-2; empty falls back to `[invoice] default_country`.
     * @nullable
     */
  practice_country?: string | null;
  /**
     * Shown in the invoice footer. The sender address of outgoing mail is operator
     * configuration and lives in `config.toml`.
     */
  email: string;
  iban: string;
  bic: string;
  bank_name: string;
  ustid: string;
  /** @nullable */
  logo_attachment_id?: number | null;
  /** Every outgoing invoice email is copied to these addresses. */
  cc_emails: string[];
  bcc_emails: string[];
}